# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Image Watermark API",
    "description": "Professional image watermarking API with customizable timestamp, address, and secure authentication.\n\n## Features\n- ✅ Auto watermark with date, time, and location\n- 🕒 Custom timestamp support (user can set any date/time)\n- 📍 Custom address support with auto text wrapping (up to 5 lines)\n- 🎨 Professional design with responsive typography\n- 🚀 High-performance processing (no disk storage)\n- 🔒 Secure with API token authentication and rate limiting\n- 💾 Preserves original image quality and resolution\n- 🖼️ Brand logo support (auto-loaded from public folder)\n\n## Security\n- API Token Authentication (Bearer token required)\n- Rate limiting: 100 requests per 15 minutes per IP\n- Input validation and sanitization\n- No local URL access (SSRF protection)\n\n## Supported Formats\n- Input: JPG, JPEG, PNG, WebP\n- Output: Same as input (preserved)\n- Max Size: 10MB\n",
    "version": "2.1.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Upload",
      "description": "Image upload and watermarking endpoints"
    },
    {
      "name": "Health",
      "description": "API health check"
    },
    {
      "name": "Documentation",
      "description": "API documentation"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "API token authentication. Get your token from environment variable API_TOKEN."
      }
    },
    "schemas": {
      "ImageResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Image processed successfully"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "properties": {
              "image": {
                "type": "string",
                "description": "Base64 encoded image (data URL format)",
                "example": "data:image/jpeg;base64,/9j/4AAQSkZJRg..."
              },
              "size": {
                "type": "integer",
                "description": "Processed image size in bytes",
                "example": 1234567
              },
              "originalSize": {
                "type": "integer",
                "description": "Original image size in bytes",
                "example": 987654
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error description"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "meta": {
            "type": "object",
            "properties": {
              "statusCode": {
                "type": "integer",
                "example": 400
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/upload": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload image file with custom timestamp",
        "description": "Upload an image file and add professional watermark with customizable timestamp and address.\n\nThe watermark includes:\n- Custom or current date/time (Asia/Jakarta timezone)\n- Custom or default address (auto-wrapped to 5 lines max)\n- Brand logo (top-right corner)\n- Professional text with outline for visibility\n\n**Authentication Required**: Bearer token in Authorization header\n",
        "operationId": "uploadFile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (JPG, PNG, WebP - Max 10MB)"
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Custom address text (optional)",
                    "example": "Jalan Sudirman No. 123, Jakarta Pusat, DKI Jakarta"
                  },
                  "time_created": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Custom timestamp in ISO 8601 format (optional)",
                    "example": "2024-12-25T14:30:00+07:00"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "binary",
                      "json"
                    ],
                    "default": "binary",
                    "description": "Response format (binary for direct image, json for base64)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image processed successfully",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "image/jpeg"
                }
              },
              "X-Original-Size": {
                "schema": {
                  "type": "integer"
                },
                "description": "Original image size in bytes"
              },
              "X-Processed-Size": {
                "schema": {
                  "type": "integer"
                },
                "description": "Processed image size in bytes"
              }
            },
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid file, size, format, or timestamp)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing or invalid API token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/upload-url": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload image from URL with custom timestamp",
        "description": "Download an image from a URL and add professional watermark with customizable timestamp.\n\n**Authentication Required**: Bearer token in Authorization header\n\nNote: Local URLs (localhost, 127.0.0.1) are blocked for security (SSRF protection).\n",
        "operationId": "uploadFromUrl",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "Direct URL to image file",
                    "example": "https://example.com/images/photo.jpg"
                  },
                  "address": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Custom address text (optional)",
                    "example": "Jalan Sudirman No. 123, Jakarta Pusat"
                  },
                  "time_created": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Custom timestamp in ISO 8601 format (optional)",
                    "example": "2024-12-25T14:30:00+07:00"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "binary",
                      "json"
                    ],
                    "default": "binary",
                    "description": "Response format"
                  }
                }
              },
              "examples": {
                "withCustomTime": {
                  "summary": "With custom timestamp",
                  "value": {
                    "url": "https://example.com/photo.jpg",
                    "address": "Jakarta, Indonesia",
                    "time_created": "2024-12-25T14:30:00+07:00",
                    "format": "json"
                  }
                },
                "currentTime": {
                  "summary": "With current timestamp (default)",
                  "value": {
                    "url": "https://example.com/photo.jpg",
                    "address": "Jakarta, Indonesia",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image processed successfully",
            "headers": {
              "X-Source-URL": {
                "schema": {
                  "type": "string"
                },
                "description": "Original image URL"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ImageResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "sourceUrl": {
                              "type": "string",
                              "description": "Original image URL"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing or invalid API token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "408": {
            "description": "Request timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Health check",
        "description": "Check API server status and get system metrics (no authentication required)",
        "operationId": "healthCheck",
        "security": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Service is healthy"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "uptime": {
                          "type": "number",
                          "description": "Server uptime in seconds"
                        },
                        "memory": {
                          "type": "object"
                        },
                        "version": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api": {
      "get": {
        "tags": [
          "Documentation"
        ],
        "summary": "Get API documentation (JSON)",
        "description": "Returns comprehensive API documentation in JSON format (no authentication required)",
        "operationId": "getDocumentation",
        "security": [],
        "responses": {
          "200": {
            "description": "API documentation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}